Update structured output unit tests to use MatchQ
The previous version of these tests used SameQ, which had the problem
of being sensitive to roundoff level differences that were difficult
to avoid. It is better for structural tests (e.g. checking that a
generated TeukolskyRadialFunction has the right form) to ignore
numerical values and to add additional tests of specific numerical
values. This commit implements the first of these, but leaves adding
new tests for later.