Avoid losing precision at the turning points for generic orbit s=-2
The four-velicities were previously computed by taking a square root.
This led to a loss of precision near the turning points, where we were
taking the square root of a number that was close to zero.

This change also moves most of the TeukolskySource code into the
convolution function. This was necessary as not enough information
was provided to the source functions to compute the Mino four-velocities
efficiently.